pyproject.toml
[tool.poetry]
name = "local-llm-docker-hosting"
version = "0.1.0"
description = ""
authors = ["Ignacy Daszkiewicz <ignacy.daszkiewicz@gmail.com>"]
readme = "README.md"
packages = [{include = "local_llm_docker_hosting"}]

[tool.poetry.dependencies]
python = "3.11.x"
sqlalchemy = "1.4.50"
pendulum = "2.1.2"
langchain = {extras = ["all"], version = "^0.0.352"}
openai = "^1.6.1"
scikit-learn = "^1.3.2"
pandas = "^2.1.4"
pillow = "^10.1.0"
torch = "^2.1.2"
torchvision = "^0.16.2"
torchaudio = "^2.1.2"
apache-airflow = {version = "2.7.3", extras = ["celery"]}
apache-airflow-providers-postgres = "^5.9.0"
apache-airflow-providers-docker = "3.9.1"
transformers = "^4.36.2"
datasets = "^2.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
